+++ src/sourcemaps/utils/sdk-version.ts
@@ -19,16 +19,24 @@ const MINIMUM_DEBUG_ID_SDK_VERSION = '7.47.0';
 // that we actually detect the "top-level" SDK first.
 const SENTRY_SDK_PACKAGE_NAMES = [
   // SDKs using other framework SDKs need to be checked first
+  '@sentry/astro',
   '@sentry/gatsby',
   '@sentry/nextjs',
+  '@sentry/nuxt',
   '@sentry/remix',
+  '@sentry/solidstart',
   '@sentry/sveltekit',
 
   // Framework SDKs
   '@sentry/angular',
   '@sentry/angular-ivy',
+  '@sentry/aws-serverless',
+  '@sentry/bun',
   '@sentry/ember',
+  '@sentry/google-cloud-serverless',
+  '@sentry/nestjs',
   '@sentry/react',
+  '@sentry/solid',
   '@sentry/svelte',
   '@sentry/vue',
   '@sentry/serverless',
@@ -36,6 +44,7 @@ const SENTRY_SDK_PACKAGE_NAMES = [
   // Base SDKs
   '@sentry/browser',
   '@sentry/node',
+  '@sentry/deno',
 ];
 
 /**