+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 ## Unreleased
 
 - feat: Only format changed files with Prettier (#670)
+- ref(remix): Use recursive true in error example creation ([#681](https://github.com/getsentry/sentry-wizard/pull/681))
+- ref(remix): Fix creation sentry example when no routes folder ([#680](https://github.com/getsentry/sentry-wizard/pull/680))
+
+## 3.32.0
+
+- feat: Add `--saas` CLI arg to skip self-hosted or SaaS selection step (#678)
+- ref: Add `--project` and `--org` args to help message and update readme (#679)
+
+## 3.31.0
+
+- fix(sveltekit): Create bundler plugin env file instead of sentryclirc (#675)
+- fix(check-sdk-version): update sentry sdk packages (#676)
+- feat(telemetry): Add telemetry for org and project CLI argument usage (#677)
 
 ## 3.30.0
 
@@ -33,11 +46,14 @@ Work in this release contributed by @MaximAL. Thank you for your contributions!
 
 ## 3.26.0
 
-- fix(nextjs): Don't add '.env.sentry-build-plugin' to .gitignore if it's already there (#610)
+- fix(nextjs): Don't add '.env.sentry-build-plugin' to .gitignore if it's
+  already there (#610)
 - feat(nextjs): Support all `next.config` file types (#630)
-- fix(nextjs): Update instrumentation and example creation logic for app or pages usage (#629)
+- fix(nextjs): Update instrumentation and example creation logic for app or
+  pages usage (#629)
 - feat(nextjs): Prompt for `reactComponentAnnotation` (#634)
-- fix(nextjs): Add missing Error.getInitialProps calls in Next.js error page snippets (#632)
+- fix(nextjs): Add missing Error.getInitialProps calls in Next.js error page
+  snippets (#632)
 - fix/feat: Improve error logging for package installation (#635)
 - fix: Properly close open handles (#638)
 
@@ -55,7 +71,8 @@ Work in this release contributed by @MaximAL. Thank you for your contributions!
 
 ## 3.24.1
 
-- fix(nextjs): Add trailing comma to `sentryUrl` option in `withSentryConfig` template (#601)
+- fix(nextjs): Add trailing comma to `sentryUrl` option in `withSentryConfig`
+  template (#601)
 
 ## 3.24.0